+++ lib/rails-bootstrap-helpers.rb
@@ -4,6 +4,7 @@ module RailsBootstrapHelpers
   module Renderers
     autoload :AbstractButtonRenderer, "rails-bootstrap-helpers/renderers/abstract_button_renderer"
     autoload :ButtonRenderer, "rails-bootstrap-helpers/renderers/button_renderer"
+    autoload :ContentTagRenderer, "rails-bootstrap-helpers/renderers/content_tag_renderer"
     autoload :Renderer, "rails-bootstrap-helpers/renderers/renderer"
   end
 
@@ -14,6 +15,7 @@ module RailsBootstrapHelpers
     autoload :FormTagHelper, "rails-bootstrap-helpers/helpers/form_tag_helper"
     autoload :LabelHelper, "rails-bootstrap-helpers/helpers/label_helper"
     autoload :OptionsHelper, "rails-bootstrap-helpers/helpers/options_helper"
+    autoload :TagHelper, "rails-bootstrap-helpers/helpers/tag_helper"
   end
 end