+++ changelog.md
@@ -6,6 +6,7 @@
 * Renamed option "placement" to "position" for `bs_popover_button`
 * Add new form tag helper `bs_submit_tag`
 * Add new button `bs_collapsible_button`
+* Add new tag helper `bs_content_tag`
 
 ### Bugs Fixed
 * "class" option not being passed through button helpers